% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_regression_preds}
\alias{check_regression_preds}
\title{Test number of level used for predictions, warn if regression was chosen for 
a classification}
\usage{
check_regression_preds(model, X)
}
\arguments{
\item{model}{model object to be used for prediction}

\item{X}{matrix to be used in predictions}
}
\description{
Test number of level used for predictions, warn if regression was chosen for 
a classification
}
