% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_data.R
\name{get_community_data}
\alias{get_community_data}
\title{Get Communities and Crime Unnormalized Data Set from the UCI respository}
\usage{
get_community_data()
}
\description{
Get Communities and Crime Unnormalized Data Set from the UCI respository
}
