% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_effects.R
\name{partial_effects}
\alias{partial_effects}
\title{Diagnostic plot for assessing effects of individual features on a principal 
component}
\usage{
partial_effects(model, X, pc_num = 1, samp_size = 20, num_load = 5,
  type = "regression")
}
\arguments{
\item{model}{A fitted model object to visualize}

\item{X}{Design matrix that the object was trained on}

\item{pc_num}{Principal component to impute}

\item{samp_size}{Number of steps to take in PC direction}

\item{num_load}{Number of top loadings to plot}

\item{type}{Outcome type, either 'classification' (binary) or 'regression'}
}
\description{
Diagnostic plot for assessing effects of individual features on a principal 
component
}
