% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_featvis.R
\name{plot.featvis}
\alias{plot.featvis}
\title{Generate a partial depedence plot for a specified feature}
\usage{
\method{plot}{featvis}(x, rug = TRUE, ...)
}
\arguments{
\item{x}{A featvis object returned from a call to totalvis}

\item{rug}{Adds a rug representation of the feature}

\item{...}{Additional optional arguments to be passed to plot}
}
\description{
Generate a partial depedence plot for a specified feature
}
