% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pinvis.R
\name{plot.pinvis}
\alias{plot.pinvis}
\title{Generate a total effect plot for a specific principal component and feature}
\usage{
\method{plot}{pinvis}(x, rug = TRUE, ...)
}
\arguments{
\item{x}{A total_vis object returned from a call to total_vis}

\item{rug}{Adds a rug representation of the principal component}

\item{...}{Additional optional arguments to be passed to plot}
}
\description{
Generate a total effect plot for a specific principal component and feature
}
