% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_totalice.R
\name{plot.totalice}
\alias{plot.totalice}
\title{Generate an ICE plots over for a principal component}
\usage{
\method{plot}{totalice}(x, num_load = 5, legend_loc = "topleft",
  center = TRUE, rug = TRUE, plot_frac = 0.1, ...)
}
\arguments{
\item{x}{A totalvis.object returned from a call to totalvis}

\item{num_load}{Number of loading to include in the legend}

\item{legend_loc}{pc_num of legend, default of topleft}

\item{center}{Logical, produced centered ice plots}

\item{rug}{Adds a rug representation of the principal component}

\item{plot_frac}{Fraction of individual conditional expectation curves to plot}

\item{...}{Additional optional arguments to be passed to plot, accepts xlab or main as arguments}
}
\description{
Generate an ICE plots over for a principal component
}
