% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_diff.R
\name{pred_diff}
\alias{pred_diff}
\title{Obtain predictions resulting from shifting a single feature}
\usage{
pred_diff(object)
}
\arguments{
\item{object}{a preddiff object passed through the partial_effects function}
}
\description{
Obtain predictions resulting from shifting a single feature
}
