% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pred_val}
\alias{pred_val}
\title{Calculate the average prediction after imputing the value for a column}
\usage{
pred_val(x)
}
\arguments{
\item{x}{A prediction object}
}
\description{
Calculate the average prediction after imputing the value for a column
}
