% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rev_pca}
\alias{rev_pca}
\title{Inverse of the PCA transformation, including centering and scaling}
\usage{
rev_pca(data, pca_object)
}
\arguments{
\item{data}{Dataframe or matrix of same dim as original data}

\item{pca_object}{A prcomp object}
}
\description{
Inverse of the PCA transformation, including centering and scaling
}
