% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Screeplot wrapper function for totalvis objects}
\usage{
scree_plot(x, main = "Scree Plot", ...)
}
\arguments{
\item{x}{An object produced from the totalvis or partial_effects functions}

\item{main}{title of the scree plot}

\item{...}{Additional arguments to be passed to screeplot}
}
\description{
Screeplot wrapper function for totalvis objects
}
