\name{totalvis-package}
\alias{totalvis-package}
\docType{package}
\title{
\packageTitle{totalvis}
}
\description{
\packageDescription{totalvis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{totalvis}
\packageIndices{totalvis}
Stuff to add
}
\author{
\packageAuthor{totalvis}

Maintainer: \packageMaintainer{totalvis}
}
