% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalvis_diag.R
\name{totalvis_diag}
\alias{totalvis_diag}
\title{Diagnostic plot for assessing effects of individual features on a prinicapl 
component}
\usage{
totalvis_diag(model, X, location = 1, samp_size = 20, num_load = 5,
  type = "regresson")
}
\arguments{
\item{model}{A fitted model object to visualize}

\item{X}{Design matrix that the object was trained on}

\item{location}{Principal component to impute}

\item{samp_size}{Number of steps to take in PC direction}

\item{num_load}{Number of top loadings to plot}

\item{type}{Outcome type, either 'classification' (binary) or 'regression'}
}
\description{
Diagnostic plot for assessing effects of individual features on a prinicapl 
component
}
