% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_featice.R
\name{plot.featice}
\alias{plot.featice}
\title{Generate an ICE plots over for a specified feature}
\usage{
\method{plot}{featice}(x, center = TRUE, rug = TRUE, plot_frac = 0.1,
  legend_cex = 1, ...)
}
\arguments{
\item{x}{A totalvis.object returned from a call to totalvis}

\item{center}{Logical, produces centered ice plots}

\item{rug}{Adds a rug representation of the principal component}

\item{plot_frac}{Fraction of individual conditional expectation curves to plot}

\item{legend_cex}{Character expansion factor for the legend}

\item{...}{Additional optional arguments to be passed to plot, accepts xlab or main as arguments}
}
\description{
Generate an ICE plots over for a specified feature
}
