% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_effects.R
\name{plot.partialvis}
\alias{plot.partialvis}
\title{Diagnostic plot for totalvis function}
\usage{
\method{plot}{partialvis}(x, legend_loc = "topleft", rug = TRUE, ...)
}
\arguments{
\item{x}{A diagvis.object returned from a call to totalvis_diag}

\item{legend_loc}{Legend placement; 'topleft', 'topright', 'bottom', ...}

\item{rug}{Adds a rug representation of the principal component}

\item{...}{Additional optional arguments to be passed to plot, accepts xlab or main as arguments}
}
\description{
Diagnostic plot for totalvis function
}
