% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_totalvis.R
\name{plot.totalvis}
\alias{plot.totalvis}
\title{Generate a partial depedence plot for a specified principal component}
\usage{
\method{plot}{totalvis}(x, num_load = 5, rug = TRUE, ...)
}
\arguments{
\item{x}{A totalvis.object returned from a call to totalvis}

\item{num_load}{Number of loading to include in the legend}

\item{rug}{Adds a rug representation of the principal component}

\item{...}{Additional optional arguments to be passed to plot}
}
\description{
Generate a partial depedence plot for a specified principal component
}
