% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_diff.R
\name{pred_diff}
\alias{pred_diff}
\title{Evaluate difference in predictions for imputing an entire PC vs
taking a step in the PC direction for a single feature}
\usage{
pred_diff(object)
}
\arguments{
\item{object}{a preddiff object passed through the partial_effects function}
}
\description{
Evaluate difference in predictions for imputing an entire PC vs
taking a step in the PC direction for a single feature
}
