% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_diff.R
\name{pred_diff}
\alias{pred_diff}
\title{Evaluate difference in predictions for imputing an entire PC vs
taking a step in the PC direction for a single feature}
\usage{
pred_diff(x)
}
\arguments{
\item{x}{Object of type "diagvis"}
}
\description{
Evaluate difference in predictions for imputing an entire PC vs
taking a step in the PC direction for a single feature
}
