% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{regression_preds}
\alias{regression_preds}
\title{Control flow for regression predictions}
\usage{
regression_preds(model, X)
}
\arguments{
\item{model}{model object to be used for prediction}

\item{X}{matrix to be used in predictions}
}
\description{
Control flow for regression predictions
}
