% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalvis.R
\name{totalvis}
\alias{totalvis}
\title{Creat an object for visiualing pdp plots for a principal component}
\usage{
totalvis(model, X, type = "regression", location = 1,
  samp_size = 100, feature = NULL, pin = NULL, ice = FALSE)
}
\arguments{
\item{model}{A fitted model object of an appropriate class}

\item{X}{Design matrix that the object was trained on}

\item{type}{Outcome type, either 'classification' (binary) or 'regression'}

\item{location}{Integer of principal component to visualize}

\item{samp_size}{Number of unique PC values marginalize and plot}

\item{feature}{Feature to plot a partial dependence plot for}

\item{pin}{Feature to pin pdp plot to over a range of a principal component}

\item{ice}{Logical, options to include individual expectation curves}
}
\value{
A pdp style plot of the specified principal component
}
\description{
Creat an object for visiualing pdp plots for a principal component
}
