% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalvis.R
\name{totalvis}
\alias{totalvis}
\title{Calculate and plot partial depedence plots for principal components}
\usage{
totalvis(model, data, type = "regression", location = 1,
  samp_size = 100, feature = NULL, pin = NULL)
}
\arguments{
\item{model}{A fitted model object of an appropriate class}

\item{data}{Dataframe of matrix of numeric type (Note: Excludes the outcome 
or the outcome is specified with'outcome')}

\item{type}{Outcome type, either 'classification' (binary) or 'regression'}

\item{location}{Integer of principal component to visualize}

\item{samp_size}{Number of unique PC values marginalize and plot}

\item{feature}{Feature to plot a partial dependence plot for}

\item{pin}{Feature to pin pdp plot to over a range of a principal component}
}
\value{
A pdp style plot of the specified principal component
}
\description{
Calculate and plot partial depedence plots for principal components
}
